Add files via upload
In this program i have tried to predict whether the game can be played or not according to the weather conditions. The model will take into account various climate-related features as input, such as temperature, humidity, wind speed, etc., and predict whether it is feasible to engage in outdoor activities like playing a sport or not.